+++ include/simgrid/kernel/routing/FloydZone.hpp
@@ -23,7 +23,7 @@ namespace routing {
  */
 class XBT_PRIVATE FloydZone : public RoutedZone {
   /* vars to compute the Floyd algorithm. */
-  std::vector<std::vector<int>> predecessor_table_;
+  std::vector<std::vector<unsigned long>> predecessor_table_;
   std::vector<std::vector<unsigned long>> cost_table_;
   std::vector<std::vector<std::unique_ptr<Route>>> link_table_;
